+++ src/styles/Icons/icons.jsx
