+++ src/styles/mermaidMuiThemeConfig.js
@@ -1,4 +1,4 @@
-import theme from '../theme'
+import theme from './theme'
 
 export default {
   palette: {
